requirements.txt
json
pytest
mysql-connector-python